v5.1.3
- (Add) Thread count to the title bar, it will show the cached managed thread count used by the program
- (Fix) Anycubic: Layer rounding problems when decoding files (#1033)
- (Upgrade) AvaloniaUI from 11.3.0 to 11.3.1
- (Upgrade) .NET from 9.0.5 to 9.0.6

+++ UVtools.Core/FileFormats/AnycubicFile.cs
@@ -2183,9 +2183,9 @@ public sealed class AnycubicFile : FileFormat
                     {
                         this[layerIndex] = new Layer((uint)layerIndex, mat, this)
                         {
-                            PositionZ = LayersDefinition.Layers.AsValueEnumerable()
+                            PositionZ = (float)LayersDefinition.Layers.AsValueEnumerable()
                                 .Where((_, i) => i <= layerIndex)
-                                .Sum(def => def.LayerHeight),
+                                .Sum(def => (decimal)def.LayerHeight)
                         };
                     }
